engine/requirements.txt
# Web framework
fastapi==0.115.12
uvicorn==0.34.0

# Database
sqlalchemy==2.0.40

# Data / HTTP
requests==2.32.3
yfinance==0.2.54
pandas==2.2.3
numpy==2.2.5

# Portfolio optimization
scipy==1.15.2
scikit-learn==1.6.1

# AI
google-generativeai==0.8.5

# Scheduler
apscheduler==3.11.0

# Config / Env
python-dotenv==1.1.0
pyyaml==6.0.2

# News
newsapi-python==0.2.7

# ETF scraping
playwright==1.51.0
